refactor: move runPurgeUsers to executor_admin_purge.go
